requirements.txt
# Browser automation
playwright==1.40.0

# Data handling
sqlalchemy==2.0.23
aiosqlite==0.19.0

# HTTP requests (for photo downloads)
aiohttp==3.9.1
aiofiles==23.2.1

# HTML parsing
beautifulsoup4==4.12.2
lxml==5.1.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0

# Data export
pandas==2.1.4

# Coordinate conversion
pyproj>=3.6.0

# Supabase
supabase==2.10.0

# Web UI
flask==3.0.0

# AI Agent Framework (Owner Enrichment)
crewai[tools]>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
litellm>=1.0.0
